requirements.txt
sentence-transformers==2.2.2
psycopg2-binary==2.9.7
requests==2.31.0
numpy==1.24.3 